Page is working, now just need to add the fwders to the main app
