+++ bin/www
@@ -1,57 +1,26 @@
 #!/usr/bin/env node
 
-/**
- * Module dependencies.
- */
-
-var app = require('../index');
+var portkbdapp = require('../app');
 var debug = require('debug')('untitled4:server');
 var http = require('http');
 
-/**
- * Get port from environment and store in Express.
- */
-
 var port = normalizePort(process.env.PORT || '8080');
-app.set('port', port);
+portkbdapp.set('port', port);
 
-/**
- * Create HTTP server.
- */
-
-var server = http.createServer(app);
-
-/**
- * Listen on provided port, on all network interfaces.
- */
+var server = http.createServer(portkbdapp);
 
 server.listen(port);
 server.on('error', onError);
 server.on('listening', onListening);
 
-/**
- * Normalize a port into a number, string, or false.
- */
-
 function normalizePort(val) {
     var port = parseInt(val, 10);
-
-    if (isNaN(port)) {
-        // named pipe
+    if (isNaN(port))
         return val;
-    }
-
-    if (port >= 0) {
-        // port number
+    if (port >= 0)
         return port;
-    }
-
     return false;
 }
-
-/**
- * Event listener for HTTP server "error" event.
- */
 
 function onError(error) {
     if (error.syscall !== 'listen') {
@@ -62,7 +31,6 @@ function onError(error) {
         ? 'Pipe ' + port
         : 'Port ' + port;
 
-    // handle specific listen errors with friendly messages
     switch (error.code) {
         case 'EACCES':
             console.error(bind + ' requires elevated privileges');
@@ -76,10 +44,6 @@ function onError(error) {
             throw error;
     }
 }
-
-/**
- * Event listener for HTTP server "listening" event.
- */
 
 function onListening() {
     var addr = server.address();
