--- index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>title</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.2/axios.min.js"></script>
-  </head>
-  <body>
-    <h1 id="message"></h1>
-    <input type="text" id="keyboard" disabled style="width:100%">
-  </body>
-  <script src="/public/javascripts/index.js"></script>
-</html>
