+++ views/layout.pug
@@ -7,5 +7,5 @@ html
   body
     block content
 
-script(src='https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.2/axios.min.js')
-script(src='/javascripts/index.js')
+    script(src='https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.2/axios.min.js')
+    script(src='/javascripts/index.js')
